code improvements, cleanup and refactor
